move inference source to lightseq dir
