+++ .gitignore
@@ -141,3 +141,4 @@ cython_debug/
 test_case/
 build/
 *.pb
+.vscode
