+++ CMakeLists.txt
@@ -47,11 +47,11 @@ if(DEBUG_MODE)
     message(STATUS "Debug computation result")
 endif()
 
-add_subdirectory(kernels)
-add_subdirectory(tools)
-add_subdirectory(proto)
-add_subdirectory(model)
-add_subdirectory(example)
+add_subdirectory(lightseq/inference/kernels)
+add_subdirectory(lightseq/inference/tools)
+add_subdirectory(lightseq/inference/proto)
+add_subdirectory(lightseq/inference/model)
+add_subdirectory(example/cpp)
 add_subdirectory(3rdparty/pybind11)
-add_subdirectory(pywrapper)
-add_subdirectory(server)
+add_subdirectory(lightseq/inference/pywrapper)
+add_subdirectory(lightseq/inference/server)
