+++ example/cpp/gpt_generation.cc.cu
