+++ example/cpp/transformer_generate_example.cc.cu
