+++ lightseq/inference/proto/CMakeLists.txt
