+++ lightseq/inference/proto/gpt.proto
