+++ lightseq/inference/proto/gpt_weight.cc
