+++ lightseq/inference/proto/transformer.proto
