+++ lightseq/inference/proto/transformer_weight.cc
