+++ lightseq/inference/server/CMakeLists.txt
