+++ lightseq/inference/server/generate_server.cc.cu
