+++ lightseq/inference/server/libserver.ldscript
