+++ lightseq/inference/server/model_config_cuda.h
