+++ lightseq/inference/tools/util.cc.cu
